third_party/blink/web_tests/platform/generic/fast/forms/text/input-select-webkit-user-select-none-expected.txt

PASS sample.selectionStart is 0
PASS sample.selectionEnd is sample.value.length
PASS successfullyParsed is true

TEST COMPLETE
